PROJECT CURATION ALF change
* float to int ALF

* hooked up fixlib to a number of sketches

* new DXF experiments

* continued V3 project curation

+++ BWSun/BWSun.pde
@@ -7,7 +7,7 @@ Boolean isFinal = true;
 //  
 //  B&W sunbeams slowly alpha out?
 //  
-float alf = 50;
+int alf = 50;
 
 //  circle vars
 float angle1 = 0, xx1, yy1, startX1, startY1, radius1 = alf;
