+++ BWSun2/BWSun2.pde
@@ -7,7 +7,7 @@ Boolean isFinal = true;
 //  
 //  B&W sunbeams slowly alpha out?
 //  
-float alf = 69;
+int alf = 69;
 
 //  circle vars
 float angle1 = 0, xx1, yy1, startX1, startY1, radius1 = alf;
