+++ EC2EVIL2012/EC2EVIL2012.pde
@@ -1,6 +1,6 @@
 Boolean isFinal = true;
 
-float alf = 13;
+int alf = 13;
 
 //  circle vars
 float angle1 = 0, xx1, yy1, startX1, startY1, radius1 = alf;
