+++ GoBig/GoBig.pde
@@ -8,7 +8,7 @@ have processing fade into the hand
 */
 Boolean isFinal = true;
 int ctMAIN = 0;
-float alf = 11;
+int alf = 11;
 
 int cX;
 int cY;
@@ -161,7 +161,7 @@ void drawCore( int XX, int YY, int maxSize ) {
   float r = 1;  // 75;
   float theta = 0;
 
-  float alf = 10;
+  int alf = 10;
   float x;
   float y;
