+++ GridLinesCirclesNOISE3/GridLinesCirclesNOISE3.pde
@@ -1,7 +1,7 @@
 //  Prozac color theme
 //
 Boolean isFinal = true;
-float alf = 11;
+int alf = 11;
 
 int cX;
 int cY;
