+++ Heart_ApacheFlexColors/Heart_ApacheFlexColors.pde
@@ -2,7 +2,7 @@
 Atomic looking PC chip maker tray
 */
 Boolean isFinal = true;
-float alf = 37;
+int alf = 37;
 float cX, cY;
 
 float ct = 0;
