+++ LSystem/LSystem.pde
@@ -1,6 +1,6 @@
 Boolean isFinal = true;
 
-float alf = 11;
+int alf = 11;
 
 int cX;
 int cY;
