+++ LionWallNoise/LionWallNoise.pde
@@ -4,7 +4,7 @@
 
 
 Boolean isFinal = true;
-float alf = 21;
+int alf = 21;
 float ii = 0;
 
 float cX, cY;
