+++ PFontList/PFontList.pde
@@ -2,7 +2,7 @@
 
 Boolean isFinal = true;
 int ct = 0;
-float alf = 37.37;
+int alf = 37.37;
 float randCirc = 10;
 
 float registerX = 0;
