+++ PatternWall3/PatternWall3.pde
@@ -2,7 +2,7 @@
 //  Rotate the circles so the crown looking opening points up
 
 Boolean isFinal = true;
-float alf = 13;
+int alf = 13;
 float ii = 0;
 
 float cX, cY;
