+++ ROTATE2/ROTATE2.pde
@@ -19,7 +19,7 @@ Specifies vertex coordinates for curves. This function may only be used between 
 Boolean isFinal = true;
 String artDailyMsg = "ERICFICKES.COM";
 //  
-float alf = 42; //37;
+int alf = 42; //37;
 
 //  circle vars
 float i, angle1 = 180, xx1, yy1, startX1, startY1, radius1 = 37;
