+++ SOL_GreenRed/SOL_GreenRed.pde
@@ -1,6 +1,6 @@
 Boolean isFinal = true;
 int ctMAIN = 0;
-float alf = 13;
+int alf = 13;
 
 int cX;
 int cY;
@@ -139,7 +139,7 @@ void drawCore( int XX, int YY, int maxSize ) {
   float r = 1;  // 75;
   float theta = 0;
 
-  float alf = 10;
+  int alf = 10;
   float x;
   float y;
