+++ SOL_SpazzyBack/SOL_SpazzyBack.pde
@@ -1,6 +1,6 @@
 Boolean isFinal = true;
 int ctMAIN = 0;
-float alf = 13;
+int alf = 13;
 
 int cX;
 int cY;
@@ -143,7 +143,7 @@ void drawCore( int XX, int YY, int maxSize ) {
   float r = 1;  // 75;
   float theta = 0;
 
-  float alf = 20;
+  int alf = 20;
   float x;
   float y;
