+++ SOL_fib2/SOL_fib2.pde
@@ -1,6 +1,6 @@
 Boolean isFinal = true;
 int ctMAIN = 0;
-float alf = 11;
+int alf = 11;
 
 int cX;
 int cY;
