+++ ShrinkingShapeGrid10/ShrinkingShapeGrid10.pde
@@ -2,7 +2,7 @@
 //  go smaller
 
 Boolean isFinal = true;
-float alf = 13;
+int alf = 13;
 
 
 int shapeSize = 10000;
