+++ colorOnTop/colorOnTop.pde
@@ -5,7 +5,7 @@
 //
 
 Boolean isFinal = true;
-float alf = 42;
+int alf = 42;
 
 int cX;
 int cY;
