+++ mcpIsh/mcpIsh.pde
@@ -3,7 +3,7 @@
 //
 Boolean isFinal = true;
 
-float alf = 37;
+int alf = 37;
 
 int cX;
 int cY;
